add order controller and namespace admin
