+++ app/views/admin/products/edit.html.erb
