+++ app/views/admin/products/new.html.erb
