+++ app/views/orders/_order_form.html.erb
@@ -1,0 +1,4 @@
+<%= form_for @order do |f| %>
+<%= f.label "Order Status" %>
+<%= f.text_field :orderstatus %>
+<% end %>
