+++ app/views/orders/edit.html.erb
@@ -1,0 +1,3 @@
+<h1>Edit Form</h1>
+
+<%= render partial: 'order_form' %>
