+++ app/views/orders/index.html.erb
@@ -1,0 +1,7 @@
+<h1>List of Orders</h1>
+
+<%= @orders.each do |order| %>
+
+<%= order.orderstatus %>
+
+<% end %>
