+++ app/views/orders/show.html.erb
@@ -1,0 +1,3 @@
+<h1>Current Order</h1>
+
+<%= @order.orderstatus %>
