+++ app/views/products/index.html.erb
@@ -1,6 +1,8 @@
 <h1>Home Page</h1>
 
-<%= link_to "Create new product", new_product_path %>
+<%if current_user.is_admin? %>
+  <%= link_to "Create new product", new_admin_product_path %>
+<% end %>
 
 <hr>
 <% @products.each do |product| %>
@@ -14,7 +16,10 @@
 <br>
 <%= product.quantity %>
 <br>
-<%= link_to "Edit", edit_product_path(product) %>
-<%= link_to "Delete", product_path(product), method: :delete, data: {confirm: "Are you sure?"} %>
+
+<%if current_user.is_admin? %>
+  <%= link_to "Edit", edit_admin_product_path(product) %>
+  <%= link_to "Delete", admin_product_path(product), method: :delete, data: {confirm: "Are you sure?"} %>
+<% end %>
 
 <% end %>
