+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
 
   devise_for :users, :controllers => { :omniauth_callbacks => "users/omniauth_callbacks" }
 
-  #resources :products, only: [:index, :show]
+  resources :products, only: [:index, :show]
   #resource :shopping_cart, except: [:new, :create]
 
   namespace :meow do
