+++ db/migrate/20170620231836_add_admin_to_users.rb
@@ -1,0 +1,5 @@
+class AddAdminToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :is_admin, :boolean, default: false
+  end
+end
