+++ spec/controllers/orders_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe OrdersController, type: :controller do
+
+end
